games-emulation/dolphin: new cpp-ipc dependency
Upstream commit: 92a201893444790d1ad1125cbb8ea92d18b6064e

Using the submodule isn't straight forward as the project wasn't
designed for this (it installed the library and headers
unconditionally). Lets go with the more straight forward solution and
package a snapshot, this means it can be patched to install correctly
and run tests.

Closes: https://bugs.gentoo.org/965908
Signed-off-by: Alfred Wingate <parona@protonmail.com>